+++ Makefile.oot
@@ -9,7 +9,8 @@ ifeq (,$(DESTDIR))
 endif
 
 obj-m += gve.o
-gve-objs := gve_main.o gve_tx.o gve_tx_dqo.o gve_rx.o gve_rx_dqo.o gve_ethtool.o gve_adminq.o gve_utils.o gve_flow_rule.o
+gve-objs := gve_main.o gve_tx.o gve_tx_dqo.o gve_rx.o gve_rx_dqo.o gve_ethtool.o gve_adminq.o gve_utils.o gve_flow_rule.o \
+	    gve_buffer_mgmt_dqo.o
 
 ifeq (,$(KERNELDIR))
 KERNELDIR := /lib/modules/$(BUILD_KERNEL)/build
@@ -21,7 +22,7 @@ default:
 
 clean:
 	@-rm -rf gve_main.o gve_tx.o gve_tx_dqo.o gve_rx.o gve_rx_dqo.o \
-	gve_ethtool.o gve_adminq.o gve_adminq_dqo.o gve_utils.o gve.o gve_flow_rule.o \
+	gve_ethtool.o gve_adminq.o gve_adminq_dqo.o gve_utils.o gve.o gve_flow_rule.o gve_buffer_mgmt_dqo.o \
 	built-in.o Module.symvers modules.order gve.ko *.mod.* .*.*o.cmd .tmp*
 
 install:
